Refactoring and addition top menu.
