+++ app/src/main/assets/cities_with_temp.json
@@ -287,6 +287,132 @@
       "id":5128581,
       "name":"New York",
       "url": "https://images.pexels.com/photos/450597/pexels-photo-450597.jpeg?auto=compress&cs=tinysrgb&h=350"
+    },
+    {
+      "coord":{
+        "lon":21.01,
+        "lat":52.23
+      },
+      "sys":{
+        "type":1,
+        "id":5374,
+        "message":0.002,
+        "country":"PL",
+        "sunrise":1540531263,
+        "sunset":1540567070
+      },
+      "weather":[
+        {
+          "id":803,
+          "main":"Clouds",
+          "description":"broken clouds",
+          "icon":"04d"
+        }
+      ],
+      "base":"stations",
+      "main":{
+        "temp":283.15,
+        "pressure":1008,
+        "humidity":71,
+        "temp_min":283.15,
+        "temp_max":283.15
+      },
+      "visibility":10000,
+      "wind":{
+        "speed":3.1,
+        "deg":250
+      },
+      "clouds":{
+        "all":75
+      },
+      "dt":1540551600,
+      "id":756135,
+      "name":"Warsaw",
+      "url": "https://www.telegraph.co.uk/content/dam/Travel/2017/July/warsaw-GettyImages-53893674.jpg?imwidth=450"
+    },
+    {
+      "coord":{
+        "lon":-75.7,
+        "lat":45.41
+      },
+      "weather":[
+        {
+          "id":800,
+          "main":"Clear",
+          "description":"clear sky",
+          "icon":"02n"
+        }
+      ],
+      "base":"stations",
+      "main":{
+        "temp":268.14,
+        "pressure":1020,
+        "humidity":73,
+        "temp_min":267.15,
+        "temp_max":269.15
+      },
+      "visibility":24140,
+      "wind":{
+        "speed":2.1,
+        "deg":50
+      },
+      "clouds":{
+        "all":5
+      },
+      "dt":1540551600,
+      "sys":{
+        "type":1,
+        "id":3694,
+        "message":0.0079,
+        "country":"CA",
+        "sunrise":1540553667,
+        "sunset":1540591093
+      },
+      "id":6094817,
+      "name":"Ottawa",
+      "url": "http://www.guidanotturna.it/wp-content/uploads/2016/10/Ottawa-1-1080x600.jpg"
+    },
+    {
+      "coord":{
+        "lon":18.06,
+        "lat":59.33
+      },
+      "weather":[
+        {
+          "id":801,
+          "main":"Clouds",
+          "description":"few clouds",
+          "icon":"02d"
+        }
+      ],
+      "base":"stations",
+      "main":{
+        "temp":277.11,
+        "pressure":1003,
+        "humidity":88,
+        "temp_min":276.15,
+        "temp_max":278.15
+      },
+      "visibility":10000,
+      "wind":{
+        "speed":2.1,
+        "deg":360
+      },
+      "clouds":{
+        "all":20
+      },
+      "dt":1540552800,
+      "sys":{
+        "type":1,
+        "id":5420,
+        "message":0.0025,
+        "country":"SE",
+        "sunrise":1540533182,
+        "sunset":1540566556
+      },
+      "id":2673730,
+      "name":"Stockholm",
+      "url": "https://www.prologis.com/sites/corporate/files/styles/2_0__md/public/images/2017/03/SE_Stockholm%2C-Sweden.jpg?h=edd508b5&itok=YuIuO-o3"
     }
   ]
 }
