+++ app/src/main/java/by/paranoidandroid/cityweather/injection/AppComponent.kt
@@ -1,8 +1,8 @@
 package by.paranoidandroid.cityweather.injection
 
 import by.paranoidandroid.cityweather.AndroidApplication
-import by.paranoidandroid.cityweather.view.fragment.CitiesFragment
-import by.paranoidandroid.cityweather.view.fragment.CityFragment
+import by.paranoidandroid.cityweather.view.fragment.base.CitiesFragment
+import by.paranoidandroid.cityweather.view.fragment.city.CityFragment
 import dagger.Component
 import javax.inject.Singleton
