+++ app/src/main/java/by/paranoidandroid/cityweather/view/fragment/base/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package by.paranoidandroid.cityweather.view.fragment
+package by.paranoidandroid.cityweather.view.fragment.base
 
 import android.support.v4.app.Fragment
