+++ app/src/main/res/values/strings.xml
@@ -10,5 +10,7 @@
     <string name="test_temperature">4 °C</string>
 
     <string name="image_transition">image_transition</string>
-    <string name="image_transition_2">image_transition 2</string>
+
+    <string name="action_add_city">Add city</string>
+    <string name="action_show_error">Show error</string>
 </resources>
